+++ src/inspector/components/JSONTree/objType.tsx
@@ -1,0 +1,15 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+/* eslint-disable @typescript-eslint/explicit-module-boundary-types */
+export default function objType(obj:any) {
+	const type = Object.prototype.toString.call(obj).slice(8, -1);
+	if (type === "Object" && typeof obj[Symbol.iterator] === "function") {
+		return "Iterable";
+	}
+
+	if (type === "Custom" && obj.constructor !== Object && obj instanceof Object) {
+		// For projects implementing objects overriding `.prototype[Symbol.toStringTag]`
+		return "Object";
+	}
+
+	return type;
+}